test(query): split the whereProperty suite and drop the benchmark figures
The suite had grown to 640 lines against the repo's ~400-line rule. Split
along the seams already in the file: the STEP-fallback behaviour, the
STEP-vs-table equivalence suite, the duck-typed store compatibility
cases, and a shared fixtures module. A pure move, no assertion changed.

Also removes hand-written benchmark figures from a test comment. Per
AGENTS.md those live only in the generated perf-numbers region; the
comment now explains why the index matters without numbers.

Corrects two statements in that comment while it is being rewritten: it
claimed the strategy test "does not fail when the filter fix is reverted"
when it does, by construction, because it asserts on which accessors were
called; and it said the fallback extracts every candidate's quantity
sets, when candidates already matched on the property side are skipped.
